+++ common/src/main/java/tocraft/remorphed/Remorphed.java
@@ -40,7 +40,7 @@ public class Remorphed {
     public static final Logger LOGGER = LoggerFactory.getLogger(Remorphed.class);
     public static final String MODID = "remorphed";
     public static final RemorphedConfig CONFIG = ConfigLoader.register(MODID);
-    public static boolean displayVariantsInMenu = true;
+    public static boolean displayVariantsInMenu = false;
     public static boolean displayTraitsInMenu = true;
     @ApiStatus.Internal
     public static final boolean foundSkinShifter = PlatformData.isModLoaded("skinshifter");