+++ CHANGELOG.md
@@ -4,6 +4,7 @@ remorphed 6.0
 - fix player blacklist handling
 - **rework menu**
 - change key to open the menu to "B"
+- don't display variants by default
 
 remorphed 5.0
 ================